+++ Modules/Intent.Modules.AzureFunctions/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.0.3
+
+- Improvement: Added "catch JsonException" in controllers to better report malformed messages.
+
 ### Version 5.0.2
 
 - Improvement: Included module help topic.